chore: reset db migration
